Refactor: 배치작업 리팩토링
- 작업 성격별로 분리하여 클래스 리팩토링
